pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai_gateway"
version = "0.1.0"
description = "Code generation provided by GitLab.com"
authors = ["GitLab AI Assisted"]
readme = "README.md"
packages = [{ include = "ai_gateway" }]

[tool.poetry.dependencies]
python = "~3.12.0"
fastapi = "^0.115.2"
dependency-injector = "^4.41.0"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
requests = "^2.32.3"
numpy = "^1.24.1"
python-dotenv = "^1.0.1"
fastapi-health = "^0.4.0"
google-cloud-aiplatform = "^1.36.4"
structlog = "^25.0.0"
asgi-correlation-id = "^4.1.0"
prometheus-fastapi-instrumentator = "^7.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
starlette-context = "^0.4.0"
google-cloud-profiler = "^4.1.0"
py-grpc-prometheus = "^0.8.0"
tree-sitter = "^0.21.0"
transformers = "^4.37.2"
snowplow-tracker = "^1.0.3"
anthropic = "^0.49.0"
pydantic = "^2.7.4"
pydantic-settings = "^2.1.0"
starlette = "^0.46.0"
httpx = "^0.27.0"
prometheus-client = "^0.21.0"
tree-sitter-languages = "^1.10.2"
jinja2 = "^3.1.3"
pyyaml = "^6.0.2"
google-cloud-discoveryengine = "^0.13.0"
litellm = "^1.35.20"
langchain = "^0.3.7"
langgraph = "^0.2.38"
langchain-anthropic = "^0.3.0"
langchainhub = "^0.1.20"
langchain-community = "^0.3.5"
boto3 = "^1.35.37"
python-gitlab = "^5.1.0"
gitpython = "^3.1.43"
unique-names-generator = "^1.0.2"
langchain-google-vertexai = "2.0.8"
gitlab-cloud-connector = { version = "^2.6.0", source = "gitlab_cloud_connector" }
sse-starlette = "^2.1.3"
q_developer_boto3 = {path = "./vendor/q_developer_boto3-1.12.0-py3-none-any.whl"}
poetry-core = "^2.0.1"
google-auth = "^2.38.0"
grpcio-tools = "^1.64.1"
grpcio = "^1.68.1"
grpcio-reflection = "^1.64.1"
sentry-sdk = "^2.17.0"

[[tool.poetry.source]]
name = "gitlab_cloud_connector"
url = "https://gitlab.com/api/v4/projects/58733651/packages/pypi/simple"
priority = "supplemental"

[[tool.poetry.source]]
name = "prompt_library"
url = "https://gitlab.com/api/v4/projects/46678122/packages/pypi/simple"
priority = "supplemental"

[tool.poetry.group.eval]
optional = true

[tool.poetry.group.eval.dependencies]
eli5 = { version = "^0.1.2", source = "prompt_library" }

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
responses = "^0.25.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
pylint = "^3.0.3"
astroid = "^3.0.2"
pytest-watcher = "^0.4.2"
pytest-randomly = "^3.15.0"
pytest-xdist = "^3.6.1"
pyfakefs = "^5.6.0"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
isort = ">=5.12.0, <6.0.0"
black = "^25.0.0"
pylint = "^3.0.3"
astroid = "^3.0.2"
mypy = "^1.9.0"
types-pyyaml = "^6.0.12.20240311"
types-protobuf = "^4.25.0.20240417"
types-requests = "^2.31.0.20240406"

[tool.pytest.ini_options]
addopts = "--disable-warnings"
testpaths = ["tests"]
filterwarnings = ["error::UserWarning"]
asyncio_default_fixture_loop_scope = "function"

[tool.poetry.scripts]
ai_gateway = "ai_gateway.main:run_app"
ai-gateway = "ai_gateway.main:run_app"
eval = "eval.main:main"
index_docs = "ai_gateway.scripts.index_docs_as_sqlite:build_indexed_docs"
troubleshoot = "ai_gateway.scripts.troubleshoot_selfhosted_installation:troubleshoot"
compute-tps = "ai_gateway.scripts.tps_test:compute_tps"
duo-workflow-service = "duo_workflow_service.server:run"

[tool.isort]
profile = "black"

[tool.pylint.main]
jobs = 0
load-plugins = [
  "lints.unsafe_dependency_resolution",
  "lints.direct_environment_variable_reference",
  "lints.file_naming_for_tests",
]

[tool.pylint.messages_control]
disable = [
  "missing-module-docstring",
  "missing-function-docstring",
  "line-too-long",
  "missing-class-docstring",
  "c-extension-no-member",
  "too-few-public-methods",
  "no-member",
  "fixme",
  "unused-argument",
  "protected-access",
  "too-many-arguments",
  "no-name-in-module",
  "too-many-locals",
  "import-error",
  "import-outside-toplevel",
  "too-many-function-args",
  "arguments-differ",
  "unspecified-encoding",
  "redefined-builtin",
  "arguments-renamed",
  "raise-missing-from",
  "too-many-ancestors",
  "abstract-method",
  "broad-exception-caught",
  "too-many-instance-attributes",
  "too-many-positional-arguments",
  "too-many-statements",
  "redefined-outer-name",
  "duplicate-code",
]
enable = [
  "unsafe-dependency-resolution",
  "direct-environment-variable-reference",
]

[tool.pylint.reports]
score = "no"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
